Nå funker dert

+++ src/main/kotlin/no/nav/helse/flex/reisetilskudd/ReisetilskuddService.kt
@@ -3,6 +3,8 @@ package no.nav.helse.flex.reisetilskudd
 import no.nav.helse.flex.db.*
 import no.nav.helse.flex.domain.Kvittering
 import no.nav.helse.flex.domain.ReisetilskuddSoknad
+import no.nav.helse.flex.domain.ReisetilskuddStatus
+import no.nav.helse.flex.domain.tilEnkel
 import no.nav.helse.flex.kafka.AivenKafkaConfig
 import no.nav.helse.flex.kafka.SykmeldingMessage
 import no.nav.helse.flex.kafka.reisetilskuddPerioder
@@ -21,18 +23,19 @@ import java.time.Instant
 @Transactional
 class ReisetilskuddService(
     private val reisetilskuddSoknadRepository: ReisetilskuddSoknadRepository,
+    private val enkelReisetilskuddSoknadRepository: EnkelReisetilskuddSoknadRepository,
     private val kvitteringRepository: KvitteringRepository,
     private val kafkaProducer: KafkaProducer<String, ReisetilskuddSoknad>,
     private val metrikk: Metrikk,
 
-    ) {
+) {
     private val log = logger()
     fun behandleSykmelding(sykmeldingMessage: SykmeldingMessage) {
         sykmeldingMessage.runCatching {
             this.reisetilskuddPerioder()
                 .splittLangeSykmeldingperioder()
                 .tidligstePeriodeFoerst()
-                .mapIndexed { idx, periode ->
+                .map { periode ->
                     ReisetilskuddSoknad(
                         sykmeldingId = sykmeldingMessage.sykmelding.id,
                         status = reisetilskuddStatus(periode.fom, periode.tom),
@@ -64,12 +67,10 @@ class ReisetilskuddService(
         }
     }
 
-
-
-
-
-    fun sendReisetilskudd(fnr: String, reisetilskuddId: String) {
-        val reisetilskudd:ReisetilskuddSoknad = TODO()//database.sendReisetilskudd(fnr, reisetilskuddId)
+    fun sendReisetilskudd(reisetilskuddSoknad: ReisetilskuddSoknad): ReisetilskuddSoknad {
+        val avbrutt = reisetilskuddSoknad.tilEnkel().copy(sendt = Instant.now(), status = ReisetilskuddStatus.SENDT)
+        enkelReisetilskuddSoknadRepository.save(avbrutt)
+        val reisetilskudd = reisetilskuddSoknadRepository.getById(reisetilskuddSoknad.id!!)
         kafkaProducer.send(
             ProducerRecord(
                 AivenKafkaConfig.topic,
@@ -79,10 +80,13 @@ class ReisetilskuddService(
         ).get()
         metrikk.SENDT_REISETILSKUDD.increment()
         log.info("Sendte reisetilskudd ${reisetilskudd.id}")
+        return reisetilskudd
     }
 
-    fun avbrytReisetilskudd(fnr: String, reisetilskuddId: String) {
-        val reisetilskudd:ReisetilskuddSoknad = TODO()//.avbrytReisetilskudd(fnr, reisetilskuddId)
+    fun avbrytReisetilskudd(reisetilskuddSoknad: ReisetilskuddSoknad): ReisetilskuddSoknad {
+        val avbrutt = reisetilskuddSoknad.tilEnkel().copy(avbrutt = Instant.now(), status = ReisetilskuddStatus.AVBRUTT)
+        enkelReisetilskuddSoknadRepository.save(avbrutt)
+        val reisetilskudd = reisetilskuddSoknadRepository.getById(reisetilskuddSoknad.id!!)
         kafkaProducer.send(
             ProducerRecord(
                 AivenKafkaConfig.topic,
@@ -93,10 +97,17 @@ class ReisetilskuddService(
         metrikk.AVBRUTT_REISETILSKUDD.increment()
 
         log.info("Avbrøt reisetilskudd ${reisetilskudd.id}")
+        return reisetilskudd
     }
 
-    fun gjenapneReisetilskudd(fnr: String, reisetilskuddId: String) {
-        val reisetilskudd:ReisetilskuddSoknad = TODO()// database.gjenapneReisetilskudd(fnr, reisetilskuddId)
+    fun gjenapneReisetilskudd(reisetilskuddSoknad: ReisetilskuddSoknad): ReisetilskuddSoknad {
+        val status = reisetilskuddStatus(reisetilskuddSoknad.fom, reisetilskuddSoknad.tom)
+
+        val gjenåpnet = reisetilskuddSoknad.tilEnkel().copy(avbrutt = null, status = status)
+        enkelReisetilskuddSoknadRepository.save(gjenåpnet)
+
+        val reisetilskudd = reisetilskuddSoknadRepository.getById(reisetilskuddSoknad.id!!)
+
         kafkaProducer.send(
             ProducerRecord(
                 AivenKafkaConfig.topic,
@@ -106,15 +117,17 @@ class ReisetilskuddService(
         ).get()
         metrikk.GJENÅPNET_REISETILSKUDD.increment()
         log.info("Gjenåpnet reisetilskudd ${reisetilskudd.id}")
+        return reisetilskudd
     }
 
-    fun lagreKvittering(kvittering: Kvittering, reisetilskuddId: String): Kvittering {
+    fun lagreKvittering(kvittering: Kvittering): Kvittering {
         return kvitteringRepository.save(kvittering)
     }
 
-    fun slettKvittering(kvitteringId: String, reisetilskuddId: String) {
-        kvitteringRepository.deleteById(kvitteringId)
-
+    fun slettKvittering(kvitteringId: String, soknad: ReisetilskuddSoknad) {
+        if (soknad.kvitteringer.map { it.id }.contains(kvitteringId)) {
+            kvitteringRepository.deleteById(kvitteringId)
+        }
     }
 
     fun hentReisetilskuddene(fnr: String): List<ReisetilskuddSoknad> {
